finish backend api.

+++ src/main/java/com/example/policy/service/MatchService.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import com.example.policy.model.AllOf;
 import com.example.policy.model.Match;
@@ -12,7 +13,7 @@ import com.example.policy.repository.AllOfRepository;
 import com.example.policy.repository.MatchRepository;
 
 import jakarta.persistence.EntityNotFoundException;
-
+@Service
 public class MatchService {
 	@Autowired
 	private MatchRepository matchRepository;
@@ -26,17 +27,26 @@ public class MatchService {
 	public Optional<Match> getMatchById(Long id) {
 	      return matchRepository.findById(id);
 	  }
-	public Match createMatch(Match Match){
-		return matchRepository.save(Match);
+	public Match createMatch(Map<String, Object> requestMap){
+		Long allOfId = Long.parseLong(requestMap.get("allOf").toString());
+		Optional<AllOf> allof = allOfRepository.findById(allOfId);
+		if (allof.isEmpty()) {
+			throw new EntityNotFoundException("AllOf not found");
+	    }
+		Match match = new Match();
+        match.setDesignater(requestMap.get("designater").toString());
+        match.setAttributeValue(requestMap.get("attributeValue").toString());
+        match.setOp(requestMap.get("op").toString());
+        match.setDataType(Integer.parseInt(requestMap.get("dataType").toString()));
+        match.setAllOf(allof.get());	
+        return matchRepository.save(match);
 	}
 	
 	public Optional<Match> updateMatch(Long id, Map<String, Object> updates) {
 		Optional<Match> optionalMatch = matchRepository.findById(id);
-
         if (!optionalMatch.isPresent()) {
             return Optional.empty();  // Match 不存在，返回空的 Optional
         }
-
         Match match = optionalMatch.get();	
 
         // 根據請求的數據更新 Match 的屬性
@@ -52,7 +62,7 @@ public class MatchService {
                     match.setOp((String) value);
                     break;
                 case "dataType":
-                    match.setDataType(Match.DataType.valueOf((String) value));
+                	match.setDataType(Integer.parseInt(updates.get("dataType").toString()));
                     break;
                 case "allOf":
                     Long allOfId = ((Number) value).longValue();
@@ -67,7 +77,17 @@ public class MatchService {
         return Optional.of(matchRepository.save(match));
     }
 	public void deleteMatch(Long id) {
-		matchRepository.deleteById(id);
+		Optional<Match> matchOptional = matchRepository.findById(id);
+		if(matchOptional.isPresent()) {
+			Match match = matchOptional.get();
+			AllOf allof = match.getAllOf();
+			List<Match> matchList = allof.getMatches();
+			matchList.remove(match);
+			matchRepository.deleteById(id);
+		}else {
+	        throw new EntityNotFoundException("Match not found with id: " + id);
+	    }
+		
 	}
 	
 }
